Address deprecation warning in log_in.
Formerly, #log_in worked in 2 stages:

  1. Make an initial login request which fails due to lack of a token.
     The response had result of 'NeedToken' and included a token.
  2. Re-send the login request using the token obtained in step 1.

When using this gem to log in to MediaWiki 1.34, I noticed that MediaWiki's
response to the first (token-less) request included the following warning:

  Fetching a token via "action=login" is deprecated.
  Use "action=query&meta=tokens&type=login" instead.

To address this warning, I've updated #log_in to always make a request for a
token before attempting to log in. (The total number of requests is unchanged.)

The #log_in method continues to accept a `token` parameter, which is used if
supplied. This may not necessary, since I suspect the `token` parameter was only
used by this class itself (during second request), but I've opted for this
approach to avoid a breaking change.

Change-Id: Ib6f475096dc681328e789d0416e57e9c798aa6f4